Subtract Value on Raw Product Computing Method refs: https://github.com/metasfresh/metasfresh/issues/18004
