+++ backend/de.metas.contracts/src/main/java/de/metas/contracts/modular/computing/purchasecontract/sales/raw/RawSalesComputingMethod.java
@@ -22,109 +22,21 @@
 
 package de.metas.contracts.modular.computing.purchasecontract.sales.raw;
 
-import de.metas.contracts.FlatrateTermId;
 import de.metas.contracts.modular.ComputingMethodType;
 import de.metas.contracts.modular.ModularContractProvider;
+import de.metas.contracts.modular.computing.AbstractRawComputingMethod;
 import de.metas.contracts.modular.computing.ComputingMethodService;
-import de.metas.contracts.modular.computing.ComputingRequest;
-import de.metas.contracts.modular.computing.ComputingResponse;
-import de.metas.contracts.modular.computing.IComputingMethodHandler;
-import de.metas.contracts.modular.log.LogEntryContractType;
-import de.metas.contracts.modular.log.ModularContractLogEntriesList;
-import de.metas.contracts.modular.settings.ModularContractSettings;
-import de.metas.inout.IInOutBL;
-import de.metas.inout.IInOutDAO;
-import de.metas.inout.InOutId;
-import de.metas.inout.InOutLineId;
-import de.metas.product.IProductBL;
-import de.metas.product.ProductId;
-import de.metas.product.ProductPrice;
-import de.metas.uom.UomId;
-import de.metas.util.Services;
 import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import org.adempiere.util.lang.impl.TableRecordReference;
-import org.compiere.model.I_M_InOut;
-import org.compiere.model.I_M_InOutLine;
 import org.springframework.stereotype.Component;
 
-import java.util.stream.Stream;
-
 @Component
-@RequiredArgsConstructor
-public class RawSalesComputingMethod implements IComputingMethodHandler
+public class RawSalesComputingMethod extends AbstractRawComputingMethod
 {
-	@NonNull private final IInOutDAO inoutDao = Services.get(IInOutDAO.class);
-	@NonNull private final IInOutBL inOutBL = Services.get(IInOutBL.class);
-	@NonNull private final IProductBL productBL = Services.get(IProductBL.class);
 
-	@NonNull private final ModularContractProvider contractProvider;
-	@NonNull private final ComputingMethodService computingMethodService;
-
-	@Override
-	public boolean applies(final @NonNull TableRecordReference recordRef, @NonNull final LogEntryContractType logEntryContractType)
+	public RawSalesComputingMethod(
+			@NonNull final ModularContractProvider contractProvider,
+			@NonNull final ComputingMethodService computingMethodService)
 	{
-		if (!logEntryContractType.isModularContractType()
-				|| !recordRef.getTableName().equals(I_M_InOutLine.Table_Name))
-		{
-			return false;
-		}
-
-		final I_M_InOutLine inOutLineRecord = getReceiptLine(recordRef);
-		final I_M_InOut inOutRecord = inoutDao.getById(InOutId.ofRepoId(inOutLineRecord.getM_InOut_ID()));
-
-		//noinspection RedundantIfStatement
-		if (inOutRecord.isSOTrx() || inOutBL.isReversal(inOutRecord))
-		{
-			return false;
-		}
-
-		return true;
+		super(contractProvider, computingMethodService, ComputingMethodType.SalesOnRawProduct);
 	}
-
-	private I_M_InOutLine getReceiptLine(final @NonNull TableRecordReference recordRef)
-	{
-		final InOutLineId receiptLineId = recordRef.getIdAssumingTableName(I_M_InOutLine.Table_Name, InOutLineId::ofRepoId);
-		return inoutDao.getLineByIdInTrx(receiptLineId);
-	}
-
-	@Override
-	public @NonNull Stream<FlatrateTermId> streamContractIds(final @NonNull TableRecordReference recordRef)
-	{
-		return contractProvider.streamModularPurchaseContractsForReceiptLine(InOutLineId.ofRepoId(recordRef.getRecord_ID()));
-	}
-
-	@Override
-	public boolean isApplicableForSettings(final @NonNull TableRecordReference recordRef, final @NonNull ModularContractSettings settings)
-	{
-		final I_M_InOutLine receiptLine = getReceiptLine(recordRef);
-		final ProductId receivedProductId = ProductId.ofRepoId(receiptLine.getM_Product_ID());
-		return ProductId.equals(receivedProductId, settings.getRawProductId());
-	}
-
-	@Override
-	public @NonNull ComputingMethodType getComputingMethodType()
-	{
-		return ComputingMethodType.SalesOnRawProduct;
-	}
-
-	@Override
-	public @NonNull ComputingResponse compute(final @NonNull ComputingRequest request)
-	{
-		final ModularContractLogEntriesList logs = computingMethodService.retrieveLogsForCalculation(request);
-		if (logs.isEmpty())
-		{
-			return computingMethodService.toZeroResponse(request);
-		}
-
-		final ProductPrice price = logs.getUniqueProductPriceOrErrorNotNull();
-		final UomId stockUOMId = productBL.getStockUOMId(request.getProductId());
-
-		return ComputingResponse.builder()
-				.ids(logs.getIds())
-				.price(computingMethodService.productPriceToUOM(price, stockUOMId))
-				.qty(computingMethodService.getQtySumInStockUOM(logs))
-				.build();
-	}
-
 }
