+++ backend/de.metas.contracts/src/main/java/de/metas/contracts/modular/computing/purchasecontract/sales/raw/RawSalesMaterialReceiptLineLog.java
@@ -25,22 +25,17 @@ package de.metas.contracts.modular.computing.purchasecontract.sales.raw;
 import de.metas.contracts.modular.ModularContractService;
 import de.metas.contracts.modular.invgroup.interceptor.ModCntrInvoicingGroupRepository;
 import de.metas.contracts.modular.workpackage.impl.AbstractMaterialReceiptLogHandler;
-import lombok.Getter;
 import lombok.NonNull;
 import org.springframework.stereotype.Component;
 
 @Component
-@Getter
 public class RawSalesMaterialReceiptLineLog extends AbstractMaterialReceiptLogHandler
 {
-	@NonNull private final RawSalesComputingMethod computingMethod;
-
 	public RawSalesMaterialReceiptLineLog(
 			@NonNull final ModCntrInvoicingGroupRepository modCntrInvoicingGroupRepository,
 			@NonNull final RawSalesComputingMethod computingMethod,
 			@NonNull final ModularContractService modularContractService)
 	{
-		super(modCntrInvoicingGroupRepository, modularContractService);
-		this.computingMethod = computingMethod;
+		super(modCntrInvoicingGroupRepository, modularContractService, computingMethod);
 	}
 }
