+++ backend/de.metas.contracts/src/main/java/de/metas/contracts/modular/computing/purchasecontract/subtractedvalue/rawproduct/SVRawMaterialReceiptLineLog.java
@@ -20,22 +20,31 @@
  * #L%
  */
 
-package de.metas.contracts.modular.computing.purchasecontract.addedvalue.rawproduct;
+package de.metas.contracts.modular.computing.purchasecontract.subtractedvalue.rawproduct;
 
 import de.metas.contracts.modular.ModularContractService;
 import de.metas.contracts.modular.invgroup.interceptor.ModCntrInvoicingGroupRepository;
 import de.metas.contracts.modular.workpackage.impl.AbstractMaterialReceiptLogHandler;
+import de.metas.product.ProductPrice;
 import lombok.NonNull;
 import org.springframework.stereotype.Component;
 
 @Component
-public class AVRawMaterialReceiptLineLog extends AbstractMaterialReceiptLogHandler
+public class SVRawMaterialReceiptLineLog extends AbstractMaterialReceiptLogHandler
 {
-	public AVRawMaterialReceiptLineLog(
+	public SVRawMaterialReceiptLineLog(
 			@NonNull final ModCntrInvoicingGroupRepository modCntrInvoicingGroupRepository,
-			@NonNull final AVRawComputingMethod computingMethod,
+			@NonNull final SVRawComputingMethod computingMethod,
 			@NonNull final ModularContractService modularContractService)
 	{
 		super(modCntrInvoicingGroupRepository, modularContractService, computingMethod);
+	}
+
+	@Override
+	@NonNull
+	protected ProductPrice getPriceActual(final @NonNull CreateLogRequest request)
+	{
+		return modularContractService.getContractSpecificPrice(request.getModularContractModuleId(), request.getContractId()).negate()
+				.negateIf(request.isCostsType());
 	}
 }
